% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edegR_function.R
\name{call.edgeR.OneWay.general}
\alias{call.edgeR.OneWay.general}
\title{Wrapper for edgeR one-way Anova with tmm}
\usage{
call.edgeR.OneWay.general(ct, Samples, total = FALSE, totalNF = NA,
  EstCounts = FALSE, EffLen = NA, totalOFF = NA, LogRaTrim = 0.499,
  SumTrim = 0)
}
\arguments{
\item{ct}{is a matrix or a data.frame with counts}

\item{Samples}{data.frame describing the groups -> 'Group'}

\item{total}{boolean}

\item{totalNF}{normalisation factor}

\item{EstCounts}{TRUE for Salmon and Kallisto}

\item{EffLen}{required for Salmon and Kallisto}

\item{totalOFF}{boolean}

\item{LogRaTrim}{parameter for calcNormFactors}

\item{SumTrim}{parameter for calcNormFactors}
}
\value{
a \code{list} with res containg the log2FC and MYlog2FC, scalingFactors is a \code{vector} for normalizing the counts, countsNormalized is a \code{matrix} with the normalized 'counts'
}
\description{
Run edgeR and tmm normalization adaptation of RBPs package !!!!
}
\author{
Claus Weinholdt
}
