% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VennFunction.R
\name{f.input.list.All.subVenn}
\alias{f.input.list.All.subVenn}
\title{Venndiagramm wrapper to plot all sub venn diagram}
\usage{
f.input.list.All.subVenn(inList, VennPlot = TRUE, VennOut = FALSE)
}
\arguments{
\item{inList}{is list}

\item{VennPlot}{is boolean to enable venn plot}

\item{VennOut}{is boolean to enable venn plot}
}
\description{
Venndiagramm for 2-Set
}
\author{
Claus Weinholdt
}
