% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VennFunction.R
\name{f.input.list}
\alias{f.input.list}
\title{Venndiagramm Whrapper}
\usage{
f.input.list(inList, VennPlot = TRUE, VennOut = FALSE)
}
\arguments{
\item{inList}{is list}

\item{VennPlot}{is boolean to enable venn plot}

\item{VennOut}{is boolean to enable venn plot}

\item{name}{vector with names of sets}
}
\description{
Venndiagramm for 2-Set
}
\author{
Claus Weinholdt
}
