% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VennFunction.R
\name{f.input3}
\alias{f.input3}
\title{Venndiagramm 3-Set}
\usage{
f.input3(p, q, r, name = c("A", "B", "C"), plotVENN = TRUE,
  vennOut = FALSE)
}
\arguments{
\item{name}{vector with names of sets}

\item{plotVENN}{is boolean to enable venn plot}

\item{A}{is the 1. set}

\item{B}{is the 2. set}

\item{C}{is the 3. set}
}
\description{
Venndiagramm for 3-Set
}
\author{
Claus Weinholdt
}
