% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VennFunction.R
\name{f.input4}
\alias{f.input4}
\title{Venndiagramm 4-Set}
\usage{
f.input4(p, q, r, s, t, name = c("A", "B", "C", "D"), plotVENN = TRUE,
  vennOut = FALSE)
}
\description{
Venndiagramm for 4-Set
}
\author{
Claus Weinholdt
}
