% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VennFunction.R
\name{f.input5}
\alias{f.input5}
\title{Venndiagramm 5-Set}
\usage{
f.input5(l1, l2, l3, l4, l5, name = c("A", "B", "C", "D", "E"),
  plotVENN = TRUE, vennOut = FALSE)
}
\description{
Venndiagramm for 5-Set
}
\author{
Claus Weinholdt
}
