% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{file_ext}
\alias{file_ext}
\title{file extention}
\usage{
file_ext(f_name)
}
\arguments{
\item{f_name}{string with filename}
}
\value{
string with file ending
}
\description{
Get file name extention
}
\author{
Claus Weinholdt
}
