% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edegR_function.R
\name{geoMean}
\alias{geoMean}
\title{geoMean of a vector}
\usage{
geoMean(ctrow, pcount = 0.25)
}
\arguments{
\item{ctrow}{is a matrix}

\item{pcount}{is pseudocount}
}
\description{
...
}
\author{
Claus Weinholdt
}
