% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VennFunction.R
\name{intersect3}
\alias{intersect3}
\title{intersect 3-Set}
\usage{
intersect3(A, B, C)
}
\arguments{
\item{A}{is the 1. set}

\item{B}{is the 2. set}

\item{C}{is the 3. set}
}
\description{
intersect for 3-Set
}
\author{
Claus Weinholdt
}
