% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VennFunction.R
\name{intersect4}
\alias{intersect4}
\title{intersect 4-Set}
\usage{
intersect4(A, B, C, D)
}
\arguments{
\item{A}{is the 1. set}

\item{B}{is the 2. set}

\item{C}{is the 3. set}

\item{D}{is the 4. set}
}
\description{
intersect for 4-Set
}
\author{
Claus Weinholdt
}
