% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VennFunction.R
\name{intersect5}
\alias{intersect5}
\title{intersect 5-Set}
\usage{
intersect5(A, B, C, D, E)
}
\arguments{
\item{A}{is the 1. set}

\item{B}{is the 2. set}

\item{C}{is the 3. set}

\item{D}{is the 4. set}

\item{E}{is the 5. set}
}
\description{
intersect for 5-Set
}
\author{
Claus Weinholdt
}
