% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edegR_function.R
\name{make.edgeR.total}
\alias{make.edgeR.total}
\title{make tolal Normalisation factor for edgeR}
\usage{
make.edgeR.total(ct, EstCounts = FALSE, EffLen = NA, LogRaTrim = 0.499,
  SumTrim = 0)
}
\arguments{
\item{ct}{is a matrix or a data.frame with counts}

\item{EstCounts}{TRUE for Salmon and Kallisto}

\item{EffLen}{required for Salmon and Kallisto}
}
\value{
a \code{list}
}
\author{
Claus Weinholdt
}
