% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edegR_function.R
\name{make.MYlog2FC}
\alias{make.MYlog2FC}
\title{Calculate geomean and log2 foldchange for two conditions}
\usage{
make.MYlog2FC(ct, pscount = 0.25, group, logIt = TRUE)
}
\arguments{
\item{ct}{is a matrix or a data.frame with counts}

\item{pscount}{is a pseudo count needed for geomean}

\item{group}{vector describing which columns should be summarized}

\item{logIt}{if TRUE ct will be log2}
}
\value{
a \code{list} with \code{matrix} of geomean values per row and \code{vector} of log2 foldchanges
}
\description{
Calculate geomean and log2 foldchange for two conditions
}
\author{
Claus Weinholdt
}
