% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{MinMeanFilterRep}
\alias{MinMeanFilterRep}
\title{row filter replicatewise}
\usage{
MinMeanFilterRep(exp_mat, groups, MinReads = 5, MeanReads = 5)
}
\arguments{
\item{exp_mat}{is a matrix or a data.frame, in which expression of genes are described by rows}

\item{groups}{vector describing which columns should be summarized}

\item{MinReads}{the minimum mean number of reads per replicate in 50% of all replicates of a group}

\item{MeanReads}{the mean mean number of reads per group of replicate}

\item{row}{this is a vector of numerics that will be passedin}
}
\value{
groupBasic and groupMean as logical \code{matrix}
}
\description{
A basic filter to be used.
}
\note{
2016-02-08
}
\author{
Claus Weinholdt
}
