% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edegR_function.R
\name{MSG}
\alias{MSG}
\title{write out message}
\usage{
MSG(..., nl = TRUE)
}
\arguments{
\item{nl}{boolean for new line}
}
\description{
...
}
\note{
https://github.com/pachterlab/sleuth/blob/048f0551a31c4aee6e59b75c86cab46ae1b3ca3a/R/misc.R#L70
}
\author{
Claus Weinholdt
}
