% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{plot.colored.dendrogram}
\alias{plot.colored.dendrogram}
\alias{plotcoloreddendrogram}
\title{Plot dendrogram colored by sample definition}
\usage{
\method{plot}{colored.dendrogram}(exp_mat, groups, method = "pearson",
  link = "average", leaf.labels = NULL, do.legend = F,
  unClusteredBranchColor = "black", legend.pos = "topright", ...)
}
\arguments{
\item{dt}{is data.table of integer}

\item{rowN}{is a vector for the new rownames}

\item{colN}{is a vector for the new colnames}
}
\value{
a \code{matrix}
}
\description{
Doe a hierarchical clustering of the data and draws a dendrogam with colored branches and leaved based on the sample definition.
}
\author{
Alexander Gabel
}
