% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{plotPCA}
\alias{plotPCA}
\title{PCA or MDS plot}
\usage{
plotPCA(exp_mat, groups, do.legend = T, log = T, do.MDS = F,
  do.ggplot = T, epsilon = 1, plot_label = F, plot_chull = TRUE,
  plot_geom_path = FALSE, plot_title = "")
}
\arguments{
\item{exp_mat}{expression matrix containing expression values}

\item{groups}{vector to define the groups of samples}

\item{do.legend}{boolean value defines for R graphic if legend should be plotted (default: TRUE)}

\item{log}{boolean defines if expression values should be logarithmized before doing PCA or MDS (default: TRUE)}

\item{do.MDS}{boolean value defines if MDS shall be calculated otherwise PCA plot will be generared (default: FALSE)}

\item{do.ggplot}{boolean value defines if ggplot2 object shall be created (default: TRUE)}

\item{epsilon}{numeric value which should be added to expression matrix as pseudo count, especially important when logarithmizing the data (default: epsilon = 1)}

\item{plot_label}{boolean value defines if samples names (column names of expression values) should be plotted}

\item{plot_chull}{boolean value defines if hull is plotted}

\item{plot_geom_path}{boolean value defines if path is plotted ; plot_chull automatic FALSE}

\item{plot_title}{string defines the plot tile}
}
\value{
a \code{list} with either a ggplot2 object (if do.ggplot == TRUE) and a data frame containing coordinates, groups, and colors
}
\description{
Generate PCA or MDS plot as ggplot2 or normal R graphic
}
\details{
Generate PCA or MDS plot as ggplot2 or normal R graphic based on a expression matrix and a group vector.
}
\examples{
\dontrun{
# here dont run example

}
}
\author{
Alexander Gabel
}
