% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edegR_function.R
\name{transform_y.counts_to_tmm.counts}
\alias{transform_y.counts_to_tmm.counts}
\title{transform y$counts to tmm.counts}
\usage{
transform_y.counts_to_tmm.counts(y, totalNF = NA)
}
\arguments{
\item{y}{is a edgeR object or matrix}
}
\value{
a \code{list}
}
\description{
Run edgeR and tmm normalization
}
\note{
use calcNormFactors(y,logratioTrim=0.499,sumTrim=0.0,doWeighting=T)
}
\author{
Claus Weinholdt
}
