% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VennFunction.R
\name{f.input2}
\alias{f.input2}
\title{Venndiagramm 2-Set}
\usage{
f.input2(l1, l2, name = c("A", "B"), plotVENN = TRUE, vennOut = FALSE)
}
\arguments{
\item{l1}{is the 1. set}

\item{l2}{is the 2. set}

\item{name}{vector with names of sets}

\item{plotVENN}{is boolean to enable venn plot}

\item{vennOut}{is boolean getting venn object or intersecting list}
}
\description{
Venndiagramm for 2-Set
}
\author{
Claus Weinholdt
}
