% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VennFunction.R
\name{f.input3}
\alias{f.input3}
\title{Venndiagramm 3-Set}
\usage{
f.input3(l1, l2, l3, name = c("A", "B", "C"), plotVENN = TRUE,
  vennOut = FALSE)
}
\arguments{
\item{l1}{is the 1. set}

\item{l2}{is the 2. set}

\item{l3}{is the 3. set}

\item{name}{vector with names of sets}

\item{plotVENN}{is boolean to enable venn plot}

\item{vennOut}{is boolean getting venn object or intersecting list}
}
\description{
Venndiagramm for 3-Set
}
\author{
Claus Weinholdt
}
